Turn this into a packagable source tree, add IndexBlinkyTape
This commit reorganizes the source tree with an eye towards PEP8 and
creating a package that can be uploaded to pypi. This means that in
the future people could get started with BlinkyTape by running two
commands:

$ pip install BlinkyTape
$ binary_clock -p /dev/ttyUSB0

I also added a new class, IndexBlinkyTape. This class extends BlinkyTape
by storing an array for LED colors, and providing a set_led_color()
function for setting the color of a specific LED. When show() is called
it will set the color of all LED's, ignoring any that were set with
sendPixel().

As a bonus, all code is now PEP8 compliant as measured by flake8.